func(V1): bm25

+++ tcvdb_text/src/main/java/com/tencent/tcvectordb/encoder/BaseSparseEncoder.java
@@ -5,8 +5,8 @@ import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.List;
 public interface BaseSparseEncoder {
-    public List<List<Pair<Long, Double>>> encodeTexts(List<String> texts);
-    public List<List<Pair<Long, Double>>> encodeQueries(List<String> texts);
+    public List<List<Pair<Long, Float>>> encodeTexts(List<String> texts);
+    public List<List<Pair<Long, Float>>> encodeQueries(List<String> texts);
     public void fitCorpus(List<String> texts);
     public Bm25Parameter downloadParams(String paramsFile);
     public void setParams(String paramsFile);
